Add TorchScript Backend for Neuropods (Python) [redacted]
Summary:
This diff adds a TorchScript backend for Neuropods. Since the C++ Neuropods interface currently
passes everything through to an embedded python interpreter using the PythonBridge ([redacted]), this
diff is adding TorchScript to both the C++ and Python interfaces of Neuropods. A fully C++
implementation of a TorchScript backend will also be implemented soon.

[redacted]

Reviewed By: yevgeni