Fix TF performance regression (#249)
In #239 we switched from using `session_->Run` to `session_->RunCallable`.

As part of this, we also used `MakeCallable` and `ReleaseCallable` in every inference call. Other TF functions including `Run` do caching internally for repeated calls with the same inputs (see [here](https://github.com/tensorflow/tensorflow/blob/r1.2/tensorflow/core/common_runtime/direct_session.cc#L980)), but `MakeCallable` does not.

It finds a subgraph and does optimization passes internally for every call.

To fix this, we cache handles returned by `MakeCallable` and release them in the destructor of `TensorflowNeuropodBackend`

## Context

This issue was found when I was rebasing #212. I noticed that the baselines on master were roughly 10x slower than the baselines in that PR. After some profiling, I found the problem and created this PR to fix it.

This exposes a general problem of not gating on benchmarks in CI. When we have done this in the past, CI variance caused many false positives, but it may be a good idea to have very generous thresholds to catch large regressions like this.

I'll create an issue about this with more details.