[TF] Fix regression caused by incorrect caching of callables (#253)
Fixes #252

When running a TF subgraph, we use callables. These let you run parts of the graph given inputs and a set of outputs.

Previously, it was possible for us to get a TF callable from our cache that had the same inputs and outputs as our current requested callable, but with a different order. Our caching code handled this, but the rest of the inference code didn't.

If the cache returned a callable that had the inputs in a different order than our request, it was possible for the input tensors to be passed in an incorrect order when calling `RunCallable`. A similar issue was theoretically possible with the outputs. This PR resolves both issues.

I'll follow up with another PR that streamlines the code a bit more.

### Test plan:
Before this PR, I was able to reproduce the issue using the test provided in #252
After applying the PR, the issue is resolved.

(The test from #252 is not included in this PR and will be landed separately)