[OPE] Speed up shutdown of the worker process (#261)
Previously, the OPE heartbeat thread could take up to 2 seconds to shutdown if we triggered a shutdown in the middle of a sleep. This caused the overall process shutdown to take much longer than it needed to.

This PR lets us interrupt the sleep to shutdown quickly by using a condition variable and `wait_for` instead of `this_thread::sleep_for`.