Convert DeviceType enum to a namespace + constants (#262)
This PR converts the `DeviceType` enum to a namespace + constants for `CPU` and `GPU`.

Previously it was very easy to accidentally make mistakes when checking devices. For example:

```
RuntimeOptions opts;
if (opts.visible_device == CPU)
{
...
}
```
would actually be comparing `visible_device` with `DeviceType::CPU`. This is comparing a `neuropod::NeuropodDevice` (which is an alias for `int`) with a `neuropod::DeviceType` (which is an enum).

With our current set of build flags, the compiler doesn't complain about this comparison. This leads to some issues that are tricky to track down.

This PR makes such comparisons less error prone by converting `DeviceType` to a namespace. This means that it is no longer valid to just reference `CPU` - it must either explicitly be `DeviceType::CPU` or `Device::CPU`

Build flags will also be updated in a future PR to expose more warnings at compile time.