[OPE] Fix race condition in heartbeat thread (#263)
This PR fixes two issues with the heartbeat thread. The first one is that it was possible for the `heartbeat_thread_` to launch before `mutex_` was created.

This would cause `mutex lock failed: Invalid argument` errors every once in a while.

This issue was fixed by changing the declaration order to put the `std::thread` last.

The second issue is that it was possible for the shutdown notification to be missed by the worker thread because we didn't lock the mutex before setting `send_heartbeat_` to false.