Add python bindings for out-of-process execution (#193)
This PR adds the ability to set a device when using out-of-process execution. It does this by setting `CUDA_VISIBLE_DEVICES` when launching the worker process.

We also add python bindings for OPE and use them in the tests. This means that the device selection tests run against OPE and ensure that it is correctly setting devices.