Load model configuration in NeuropodBackend (#265)
This PR moves some configuration responsibility from `Neuropod` to `NeuropodBackend`.

`Neuropod` should be a very thin layer that has the job of finding and delegating to the right backend implementation. This will be made more clear in future PRs.

This PR also enables some future changes around the ability to lazily load models (i.e. not in the constructor) and cleaner OPE initialization.