[Seal] Basic implementation of `to` operation (#344)
This PR sets up some basic infrastructure for `seal`. It introduces a `to` method on `NeuropodTensor` that can move tensors to a specified device (GPU, worker process, remote server, etc.) before `infer` is called.

This lets us do a better job of parallelizing data transfer with compute.

For tensors that are not on CPU, all the publicly visible data access methods will throw an error.

In this PR, `to` is a noop, but future PRs will introduce implementations for each backend.

Serialization of tensors on other devices will also be more concretely specified in future PRs.