[OPE] Use GPU UUIDs instead of indices for more stability (#337)
When launching a worker process for OPE, we don't want to just pass through a device index because `CUDA_VISIBLE_DEVICES` may be applied to the main process, causing an index mismatch between the worker and the main process.

For example:

If the real CUDA GPU ordering is:
```
0 -> GPU A
1 -> GPU B
2 -> GPU C
3 -> GPU D
```

and the main process is launched with CUDA_VISIBLE_DEVICES=2, it might see the following:
```
0 -> GPU C
```

If that process tries to run a model using OPE (with `visible_device` in `RuntimeOptions` set to `0`), we expect it to use `GPU C` (as that is GPU 0 in the main process).

However, if it launches a worker process setting CUDA_VISIBLE_DEVICES=opts. visible_device (as it currently does), the worker will see
```
0 -> GPU A
```

because `opts.visible_device` index is 0.

To fix this, we use GPU UUIDs instead of indices when launching an OPE worker process.