[OPE] Improve IPC serialization of structs (#362)
This PR uses reflection to serialize structs instead of requiring us to define `ipc_serialize` and `ipc_deserialize` for every struct we want to pass to the worker process.

This is much less error prone than the manual method since adding a field to a struct no longer requires modifying serialization code.

As part of this, we switch to building Neuropod with C++17. We still build in a manylinux2014 compatible sysroot so we can't use all C++17 library features, but this lets us use some C++17 language features.