[Bindings] Fix a type issue for string tensors of zero elements (#386)
### Summary:
Previously, the python bindings did not correctly set the numpy array type for string tensors with zero elements.

This PR adds a fix along with additional tests.

### Test Plan:
Added some new cases to the existing serialization tests:
1. A C++ test for string tensors with zero elements
2. A Python test for string tensors with zero elements

Before the fix in this PR, the C++ test passed, but the python one failed (implying an issue in the python bindings). With the fix, both tests pass.