Ensure consistent output hashes when packaging the same model (#394)
### Summary:

Previously, when packaging a model multiple times, the hash of the output package could change even if the input did not (e.g. all the files contained in the package had the same hashes).

This was mostly because the output Neuropod package (which is a zipfile) stored timestamps of all the included files. It was also possible that we'd add files to the zip in different orders (because `os.walk` doesn't guarantee an order).

This PR sets the timestamp of all included files to a constant and also ensures that we always add the files in the same order.

### Test Plan:

Added a test to package a model twice and compare the `sha256` of the output. Also confirmed consistent hashes on a model internally.