Fix a bug with string tensors and OPE (#433)
### Summary:

#397 describes a segfault that happens in certain cases when creating several string tensors and using OPE with multiple models across threads.

This PR implements a test that can reproduce the segfault a very high percentage of the time along with a fix for the root cause of the issue.

This line was the root cause of the issue:

https://github.com/uber/neuropod/blob/14a7e67edfe61930e123582be26075b9efcbe8f7/source/neuropod/multiprocess/shm_tensor.hh#L195-L196

We were not storing the unique_ptr created by `make_unique` so the destructor was immediately called. The reason this didn't cause visible issues in all cases has to do with how caching of shared memory blocks is implemented.

When `block` is destroyed in the code above, the underlying shared memory block is returned to an in-process cache as it's likely that the block will be loaded again. As such, the SHM block is not actually unmapped so the address returned by `get_raw_data_ptr()` still points to a valid location.

In #397, it was observed that the issue only happens when running inference in parallel across multiple threads and using the default setting of `free_memory_every_cycle = true;`. This is because such a configuration significantly increases the likelihood of clearing a shared memory cache (and unmapping the blocks in the cache) while another thread is accessing data in one of those shared memory blocks.

Fixes #397.

### Test Plan:

Added a test that consistently failed before the fix, but passes on every run after the fix.