Added logs for situations where the backend could not be loaded (#445)
### Summary:
It's not clear when the find_registered_backend method returns a null pointer, the specific reason for doing so. This change attempts to make it clear by documenting the reason why the backend isn't found.

### Test Plan:
Create a neuropod locally with required backend where we know the backend cannot be loaded. See log getting printed.