[CI] Ensure we can make requests to the GitHub API from CI (#464)
### Summary:
Previously, we'd only make requests to the GitHub API when doing release builds. In some cases we'd see failures when making https requests because of broken dependencies in CI. These failures would only be visible during release builds.

This PR fixes such an issue with Python 3.5 on macOS. It also enables CI pipelines to always make a request to the GitHub API so we can catch similar issues earlier. 

### Test Plan:

CI