Fix dylib id of libneuropod.so on macOS (#469)
### Summary:

Fixes #468

This change modifies the dylib id of `libneuropod.so` from a bazel internal path (e.g. `bazel-out/darwin-fastbuild/bin/neuropod/libneuropod.so`) to a generic one (`@rpath/libneuropod.so`).

Previously, we were modifying all the binaries that depended on `bazel-out/.../libneuropod.so` to depend on `@rpath/libneuropod.so` instead. This required logic in multiple places and more importantly didn't fix the issue for external code that links against `libneuropod.so`

By changing `libneuropod.so` itself, we can avoid changing all the binaries and libraries that depend on it.

### Test Plan:

CI