Support StringTensor as model output in Neuropod JNI (#515)
### Summary:
In neuropod JNI, it does not support STRING_TENSOR as model output. The reason is when it tries to create a Java Tensor from model output, it would check the tensor type. Only numerical tensors are supported and these tensors would be converted to ByteBuffer. However, Java NeuropodTensor itself do support STRING_TENSOR. It can get strings from the tensor data itself. Thus we refactor nativeGetBuffer method. This method would assign a null buffer to bytebuffer if it is a string tensor

### Test Plan:
Unit test