Add unicode support and tests (#527)
### Summary:

This PR modifies Neuropod string handling to assume that strings are `UTF-8` encoded. This adds support for things like emoji 😎 (and all the other characters supported by unicode), but does have a few downsides and introduces some complexity.

I'm working on some docs that dig into the difficulties of string tensors and especially string tensors + unicode, but they're not super relevant for the functionality in this PR so I'll post that separately.

The major impact of this PR is the following change.

**Breaking change:** opaque sequences of bytes will no longer generally work as strings (because the python bindings now expect strings to be valid UTF-8 strings).

This can be mitigated by adding a `bytes` type to Neuropod that supports arbitrary byte sequences. Please comment on this PR or open an issue if this is an important use case for you and we can prioritize it.

### Test Plan:

CI + modified the string test to use emojis