Fix a bug with loading large models (> 2gb) (#544)
### Summary:

The zipfile library we were previously using had trouble with files larger than ~2gb (the max value of an `int32`). This broke loading of large models. This PR modifies Neuropod to use a different zip library and also adds a test to ensure we can package and load large models correctly.

### Test Plan:

New test that packages and loads a large (~3gb) TorchScript model + CI