Bump version to 0.3.0rc7 and relax backend version requirement (#552)
### Summary:

Update the Neuropod version to `0.3.0rc7`.

This PR also allows Neuropod `0.3.0rc7` to use backends from `0.3.0rc6`. This is a temporary workaround to avoid having to reinstall backends whenever upgrading the main library. This will be improved once we have a long term release/ABI compatibility strategy (along with an ability to test and enforce it in CI).

### Test Plan:

CI + Local tests by @qiyanz to check backend compatibility